requirements.txt
Django==1.5
Pillow==1.7.8
South==0.7.6
argparse==1.2.1
djangorestframework==2.2.1
google-api-python-client==1.0
httplib2==0.7.7
python-gflags==2.0
wsgiref==0.1.2
